Added Character Horizontal Movement
Added horizontal movement.
Issues to fix:
Character U-turns when at standstill.
Current Plan: add several checks, for if
1. the character is at a standstill
2. if the character's horizontal velocity is less than a threshold.

How it works:
Instead of manually changing the velocity components of the character, we instead adjust the direction they are running. We also add a cap for how far the adjustment can be per frame, which allows the movement to 'smooth out', rather than suddenly jump from one velocity to the next. This makes the controls feel heavier.