Added a greatsword swinging state.
Added a state for swinging a greatsword, and charging it. I'm going to need to make revisions to these states to allow the 'movement' to be stored in a variable other than forwardvel. Otherwise, the character will always need to look at the direction they are moving, which will prevent the lockon from working quite how we want it to.