Added Adjustable Camera
Added several classes to enable camera control. My hope is to extend 'GameCameras' to have a hierarchy of potentially active cameras. These will have priorities that can be assigned. A higher priority means that the camera overwrites lower ones, until only one is selected.

The actual player camera, for now, needs to be adjusted to have mouse control, and needs to have max 'vertical' rotation added, so that it doesn't flip the camera over the player.

Also, now there will be a print statement for if the current state is not one assigned a specialized function. That way, it will be easier for me to tell why my code is failing (though I should probably throw an error instead.)