Player Springarm Fix
Resolved an issue with the Springarm for the player camera. When moving the camera to a position, make sure that it isn't going to be inside the ground that it needs to collide with. In this case, we move it to be slightly offset from the character's position. We will later be using a form of 'focal point adjuster'.