Got rid of unnecessary constant
ACT_JUMP is now an enum, so it is unnecessary. I did find an oddity where the character seems to 'jitter' when you turn the camera at the same time that you are running. It might be due to the disconnect between physics and the process framerates, as increasing the physic ticks reduces the jitter. Maybe I can move the rotation of the character into _process, and keep the movement in physics.