Fixed Jitter in Character Turning
I mentioned before that there was a jitter in the character's movement when turning.

Upon further inspection, this was caused by a misplaced indentation. Before we rotate the character, we check against a maximum turning angle, and clamp down. However, I accidentally included the code that actually rotated the character within the if block, meaning that the character would not turn unless the turn was bigger than the max. Adding a variable for if we should rotate the player, and moving the code to rotate the player into a separate block fixed the bug.