Added Coyote Time
Added coyote time to our character. Coyote time is a reference to looney tunes, where Wile E has some time before gravity begins to act upon him. In this case, we've made it so that there is a timer that determines if the character should transition to the 'jump' state. Of course, since we never gave the character an upward velocity in this case, this makes the character start falling.

The current coyote time is 60 frames, which is obviously excessive, but I did this to make it clear to myself whether the code was working.

currently, there isn't a coyote timer check in the greatsword walking code, and we do need to decide what to do if the character uses the greatsword in midair. We will put whatever code that is in greatsword check, most likely.

Also, I've added gravity to the walking states, and fixed a bug causing the character to never transition back to standing.