Added Ability to Cut Jump
I added the ability to cut a jump short. When the player lets go of the jump button, if they are in the air and have a rising velocity, the code will cause their vertical velocity to be cut short. Note that this is locked behind a boolean, which means we can control when the player has access to this feature. If a player is launched from the ground by another object, we don't want to be able to cut their jump short, especially if they are launched sky high. However, if the player jumps by themselves, then we would want this functionality. Note that this code also assumes that the brake jump and normal jump are the same state. If I make changes to separate the two, I will need to set the boolean before going into the brake jump state.