Added Horizontal Camera 'Prediction'
Added ability for camera to move horizontally, based on the relative horizontal movement of the players. Looking into Pseudoregalia's gameplay some more, it appears that sometimes this functionality turns off. It may be due to the effect only activating in trigger areas, or the effect being toggled based on properties of the surface last stepped on. Either way, I may need to make adjustments to my design to enable changing the active camera, so maybe I can also change the state of the camera follow functionality.