Added Items
Added Items which can be assigned a type and effect. These can be assigned to action buttons, which will allow the user to use their actions on demand. Next step is to finish developing a chest object (to give players Items), and to figure out how to store items in an inventory. My main concern has to do with the possibility of multiple weapons or adding items that are consumables and thus can have quantities.

Added an HP bar, which is updated whenever the player's HPChanged function is used. Next step is to make sure that this is called whenever the value of HP actually changes.

Reorganized files into folders. These should make things more organized in the future. However, it does make this update look bigger than it is.