Added Inventory Save and weapons
Player can now use weapons, and Items are now saved in your inventory. This code is probably pretty janky and will probably break, but for now it seems to be working well enough. Next precaution will probably be to modify the check on the equipping of weapons, although that will probably be simple (changing =s to is, or something similar)