Fixed weapon bug, added new Inventory ui elements
The Pause menu now has a UI that displays the items in your inventory. I will need to change the structure in the future, but this does serve as a proof of concept that I can create a menu the player can interact with.

Next step is to add the ability to assign these items to the action buttons. I'm also going to want to change the action buttons to trigger on the d pad instead, since these are buttons rarely used.

Also, there was an inconsistency where one of my weapons resources was of type 'Key' instead of weapon, so it wasn't triggering the proper escape clause in the code to become a weapon item. This is fixed now.