Added Images to Items, Player Skeleton
This commit comes with 2 big changes. First, I've added some images that can be used for items. Second, I've added a player skeleton and figured out how to use bone attachments to keep objects attached to a player model. You can use a BoneAttachment3D node, make the skeleton the parent of said node, and then keep a node as a child for adjusting the objects' positioning and angle.