Fixed player character mesh issues
I had to change the skeletonTest scene too, but I fixed an issue where some faces in the player model were inverted, leading to odd behaviors. I also made modifications so that the save holder can delete saves as well as load them. In the future, I will probably make changes so that 1. the newgame button is disabled if all save slots are filled, and 2. the continue feature is set to false IF all saves have been deleted.