Modified Lockon
Made changes to lockon in various ways. Now the lockon is dynamic, and checks amongst various targets that are in range. The lockon target must also be on screen. I might decide to modify the code further to make the judgement of 'closest' depend on other factors than raw 3-dimensional difference, and I also feel that it may be good to change the lockon rotation code at some point, because it feels a little too fast and sudden right now.