Action Item Remapping.
Made changes to add keyboard controls to the game. Also made changes to the action buttons to 1. allow you to remap items to the slots, and 2. makes them show the icon of the current action item assigned to them. This is done through the use of an 'actionItemChange' signal. Admittedly, the controls may be getting a bit busy or cumbersome for mouse and keyboard, I may need to add a button to rotate the camera 'horizontally'.