Item changes, Treasure chest revisions
Treasure chests have been implemented. Once a treasure chest is opened (by an attack), it will add its flag to the Globals node's openedChests property. These are saved to the file, so opening a save will not make chests reset to being unopened. Also, some items have been changed to be given unique ids, so that the utility items can be re-added properly.