Added Dash State, Death Boundary to testing
Added an item called the Dash crystal that gives the player a dash upon use. I also decided to add a death boundary that reset the scene upon collision. However, this caused me to learn something very important to Godot's workings.

When you tell the tree to change to another scene, it will delete nodes that were generated during the previous scene load. HOWEVER, any nodes that are generated outside of the scene load, such as projectiles, enemies, the player character, etc. will NOT be deleted. This was something I learned when my code accidentally generated 30 player characters upon a reload, because the old player character was still touching the death boundary. This has been fixed by calling a function on all members of a group called 'DestroyOnLoad'.