Fixed Dashing inconsistencies.
Dashing had an issue where it was not turning properly. I realized after the fact that I was forgetting to rotate the movement vector, which meant it was trying to orient to the wrong axis.

Also, a friend pointed out to me that the issue with the character not getting deleted can probably be remedied by making sure to add them to the root node manually, which would be a lot smarter in the long run.

Also, I changed the turning speed while dashing, and the length of the dash. This was mainly for testing, I will need to make further revisions later to make it feel nicer.