Added Ledge Grab
Ledge grab added. Dear god that took longer than it should have. Get_Collision_safe_fraction is buggy as heck and I am never using it again if I can avoid it. Instead, I used move_and_collide, as that will sop an object upon collision. A future enhancement I might need to add is some refinement to make it so that if you are in a corner, the character will decide to grab the closest edge to their front, but for now the ledge grab works fine.