Added dithering, fixed camera 'sink' issue
Note to self: In the future, label collision layers. Right now it was hard to tell why the character was grabbing onto the enemy.

Anyways, I made several changes during this commit, though they are mostly minor modifications. The camera no longer sinks into the floor when dragged along it (this is because I added an intermediate node and nudged the camera forward a little bit). While I was at it, I changed the player to use a material that dithers when close up. I might change this in the future, as it caused the model to sort of 'invert' but for now it will do the purpose it was meant for. Also, I changed the degree to which you must pull back the stick or push it forward to initiate logic in the act_ledge_grab state. Next I might mess around with adding actions to items that activate when their respective button is held down. For example, I'd like a wall slide, but that will need to be a continuous action, and not a one-time thing. Maybe I will add a kick-jump first.