Fixed ledge grab jump cancel bug
I noticed that the player acted a little funny when it came to letting go from a grab. I believe this is because the state of 'canCutJump' wasn't being set to specifically be false when exiting the ledge grab state, and so added explicit changes.
Also, I made the velocity change optional, based on a boolean. I may later separate this into 2 or 3 booleans, for the different axes.

There were also inconsistencies in the player position and the apparent position of the platform, because the player moves in physics process, but the platform and animation were using idle. This has now been resolved.