Improved Ledgegrab (for real)
Ledgegrab has now been improved to be more accurate, AND to work for grabbing onto slopes. However, there have been sacrifices.
1. the new approach requires the player to be on the wall, rather than in front of a wall. Later I may be able to add code that performs a preliminary 'move_and_collide', and then moves the player back, but for now I will consider the extra length of grabbing a  luxury.
2. If the wall under the slope is towards the player in any way, the player will not grab. This is because of some code I added that checks if the player is too close to the ledge for a grab.

Also, the 'ground check' code currently uses a cylinder, instead of a capsule. That should probably be fixed.

Other notes:
Currently the ledge grab sometimes seems to not be recognized? need to figure out why this is, probably something to do with either the raycast or not being 'on' the wall. Although, it might be because of my code to 'let go' activating.
The ledge grab and floor check assumes a character height of 2. See about making this more generic.
The character always faces the wall even if they shouldn't ledge grab. see about fixing this by only keeping the turn if they character grabs.