Added Aerial Movement
I added some aerial movement to the main character, although I'm sill messing with it to get the feeling right. also, there was a bug where pressing 'back' would cause a character to instantly let go form a ledge grab, even if it was just from rotating the camera. I will need to consider the best way to check for if the player intends to let go in the future.