Reintroduced ledge grab movement
Added some more changes to the ledge grab (yes, really), and made it move with platforms again.

Here, the main change introduced is that the raycast now will check a little ways into the shape. This is because there is a rare case where the collision point doesn't intersect with the actual shape, which means that effort to get the grabbed object's name is null. this leads to odd consequences, like the player sinking to the floor.

I might need to modify it so it uses a consistent ray scan, but this approach at least allows a little more range to the ledge grab.