Added ground check and an incomplete wall slide state.
Fixed a few bugs that were again causing the character to get stuck in the ground, and also added a wall sliding state. Unfortunately, I did find a bug where if a corner is small enough, the character will start to fall more slowly than before. I suspect it is due to how the horizontal velocity is still sending the character into the wall, which results in some part of the move slide code shorting the vertical drop.